Fix nightly build on m68k. Why is LONG a long int on m68k?
git-svn-id: https://svn.aros.org/svn/aros/trunk@54875 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/aminet/games/think/muimine/BTWindow.c
@@ -575,12 +575,12 @@ static void FillBestTimeText(struct BTWindowData * data, struct LevelData * pLev
                 */
                 if (i == hiRank)
                 {
-                    sprintf(timestr, "\033b%d ", p->Time);
+                    sprintf(timestr, "\033b%d ", (int)p->Time);
                     sprintf(namestr, "\033b%s",  p->Name);
                 }
                 else
                 {
-                    sprintf(timestr, "%d ", p->Time);
+                    sprintf(timestr, "%d ", (int)p->Time);
                     strcpy(namestr, p->Name);
                 }
             }
